Split coloring to separate file and enable it only on isatty
